+++ src/crh/mod.rs
@@ -22,3 +22,14 @@ pub trait FixedLengthCRH {
     fn setup<R: Rng>(r: &mut R) -> Result<Self::Parameters, Error>;
     fn evaluate(parameters: &Self::Parameters, input: &[u8]) -> Result<Self::Output, Error>;
 }
+
+pub trait FixedLengthTwoToOneCRH {
+    /// The size for length input and right input
+    const INPUT_SIZE_BITS: usize;
+
+    type Output: ToBytes + Clone + Eq + core::fmt::Debug + Hash + Default;
+    type Parameters: Clone + Default;
+
+    fn setup<R: Rng>(r: &mut R) -> Result<Self::Parameters, Error>;
+    fn evaluate(parameters: &Self::Parameters, left_input: &[u8], right_input: &[u8]) -> Result<Self::Output, Error>;
+}
