+++ src/lib.rs
@@ -19,6 +19,7 @@ pub(crate) use ark_std::{borrow::ToOwned, boxed::Box, vec::Vec};
 
 pub mod commitment;
 pub mod crh;
+pub mod merkle_tree_old;
 pub mod merkle_tree;
 pub mod prf;
 pub mod signature;
